Allow cfg enabling healthchecks when db migrate is enabled
For non production scenerios it can be useful to set db migrate to
true always and yet still enable the ecs tg health check. It's very
convenient for demo, test sites etc. that may be frequently updated.

For production you would not want this because if a db migration
runs for longer than the healthceck period it will be interrupted,
and that can be bad times.